requirements.txt
# Snowflake
snowflake-connector-python>=3.7.0

# Dashboard
streamlit>=1.31.0
altair>=5.0.0
pandas>=2.0.0

# Testing
pytest>=8.0.0

# Environment
python-dotenv>=1.0.0
